+++ src/jvmTest/kotlin/com/insanusmokrassar/krontab/utils/RunTest.kt
@@ -1,0 +1,9 @@
+package com.insanusmokrassar.krontab.utils
+
+import kotlinx.coroutines.CoroutineScope
+import kotlinx.coroutines.runBlocking
+
+/**
+ * Workaround to use suspending functions in unit tests
+ */
+actual fun runTest(block: suspend (scope: CoroutineScope) -> Unit) = runBlocking(block = block)