+++ CHANGELOG.md
@@ -6,3 +6,7 @@
 * Updates in libraries:
     * Coroutines `1.3.2` -> `1.3.3`
     * Klock `1.7.3` -> `1.8.6`
+
+## 0.2.1
+
+* Added support of flows: now any `KronScheduler` can be convert to `Flow<DateTime>` using `asFlow` extension